refactor(response-preview): move getBinaryPreviewType to utils
